ds1 editor: moved ds1 states into separated file
