+++ hswidget/ds1widget/doc.go
@@ -1,0 +1,3 @@
+// Package ds1widget provides a giu.Widget implementation for viewing and editing
+// the DS1 files.
+package ds1widget
